Organize types and add doc for exposed fn
