--- express/src/utils/handler-config.ts
@@ -1,5 +1,0 @@
-export interface DecipherHandlerConfig {
-  codebase_id: string;
-  customer_id: string;
-  exclude_request_body?: boolean;
-}
